+++ src/commands/getDeleteAllConversationsCommand.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { ConversationManager } from "../Conversation/ConversationManager";
+import { ConversationManager } from "../Conversation/temp";
 import { ChatWebviewProvider } from "../providers/ChatViewProvider";
 
 export const getDeleteAllConversationsCommand = (
