+++ src/commands/getReviewCodeCommand.ts
@@ -4,7 +4,7 @@ import { SettingsProvider } from "../helpers/SettingsProvider";
 import { TextBlock } from "../ChatModel/ChatModel";
 import { sendChatMessage } from "../helpers/sendChatMessage";
 import { getCodeReference } from "../helpers/getCodeReference";
-import { ConversationManager } from "../Conversation/ConversationManager";
+import { ConversationManager } from "../Conversation/temp";
 
 /**
  * Queries the model for a reviewed, edited version of the current file contents.
